move test_decimal.py
