+++ python/cudf/cudf/tests/private_objects/test_column.py
@@ -1,5 +1,6 @@
 # Copyright (c) 2020-2025, NVIDIA CORPORATION.
 import sys
+from decimal import Decimal
 
 import cupy as cp
 import numpy as np
@@ -17,6 +18,7 @@ from cudf.core._compat import (
 )
 from cudf.core.buffer import as_buffer
 from cudf.core.column.column import _can_values_be_equal, as_column
+from cudf.core.column.decimal import Decimal32Column, Decimal64Column
 from cudf.testing import assert_eq
 from cudf.testing._utils import assert_exceptions_equal
 
@@ -585,3 +587,46 @@ def test_datetime_can_cast_safely():
     )
 
     assert sr._column.can_cast_safely(np.dtype("datetime64[ns]")) is False
+
+
+@pytest.mark.parametrize(
+    "data_",
+    [
+        [Decimal("1.1"), Decimal("2.2"), Decimal("3.3"), Decimal("4.4")],
+        [Decimal("-1.1"), Decimal("2.2"), Decimal("3.3"), Decimal("4.4")],
+        [1],
+        [-1],
+        [1, 2, 3, 4],
+        [42, 17, 41],
+        [1, 2, None, 4],
+        [None, None, None],
+        [],
+    ],
+)
+@pytest.mark.parametrize(
+    "typ_",
+    [
+        pa.decimal128(precision=4, scale=2),
+        pa.decimal128(precision=5, scale=3),
+        pa.decimal128(precision=6, scale=4),
+    ],
+)
+@pytest.mark.parametrize("col", [Decimal32Column, Decimal64Column])
+def test_round_trip_decimal_column(data_, typ_, col):
+    pa_arr = pa.array(data_, type=typ_)
+    col_32 = col.from_arrow(pa_arr)
+    assert pa_arr.equals(col_32.to_arrow())
+
+
+def test_from_arrow_max_precision_decimal64():
+    with pytest.raises(ValueError):
+        Decimal64Column.from_arrow(
+            pa.array([1, 2, 3], type=pa.decimal128(scale=0, precision=19))
+        )
+
+
+def test_from_arrow_max_precision_decimal32():
+    with pytest.raises(ValueError):
+        Decimal32Column.from_arrow(
+            pa.array([1, 2, 3], type=pa.decimal128(scale=0, precision=10))
+        )
